pyproject.toml
[project]
name = "streamlit-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
